chore(lotto): create package structure, move Lotto.java
Create three packages
- controller, domain, exception
Create class(enum) files
- controller.lottoController
- domain.Lottos, domain.UserFinance, domain.Winnings
- exception.InputException
Move Lotto.java to lotto.domain package
