+++ src/main/java/lotto/controller/LottoController.java
@@ -1,0 +1,4 @@
+package lotto.controller;
+
+public class LottoController {
+}
