+++ src/main/java/lotto/domain/Lottos.java
@@ -1,0 +1,4 @@
+package lotto.domain;
+
+public class Lottos {
+}
