+++ src/main/java/lotto/domain/UserFinance.java
@@ -1,0 +1,4 @@
+package lotto.domain;
+
+public class UserFinance {
+}
