+++ src/main/java/lotto/domain/Winnings.java
@@ -1,0 +1,4 @@
+package lotto.domain;
+
+enum Winnings {
+}
