+++ src/main/java/lotto/exception/InputException.java
@@ -1,0 +1,4 @@
+package lotto.exception;
+
+public class InputException {
+}
